# docs/api.yaml
openapi: 3.0.2
info:
    title: OGBank
    version: 1.0.0
    description: Ojarand ja Galkin pank
    contact:
        name: Ojarand Karl Galkin Hannes-Gregor
        email: hannes-gregor.galkin@khk.ee
paths:
...
    /api/transactions:
        summary: Make a transaction
        description: Make a transaction
        get:
            tags:
                - Transactions
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/successfulResponse'
                            examples:
                                successfulTransactionHistory:
                                    value:
                                        transactions:
                                            createdAt: '2020-11-03T17:23:04.056Z'
                                            status: completed
                                            _id: 5ga19248f1ff7c46a80eeb5d
                                            userId: 5ga18d06e11be12268b48733
                                            amount: 500
                                            currency: EUR
                                            accountFrom: f496138686071081018
                                            accountTo: f496138686071081017
                                            explanation: raha
                                            senderName: karlos
                                            __v: 0
                                            receiverName: hannes
                                            statusDetail: ''
                    description: Successful transaction history
                '400':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/errorResponse'
                            examples:
                                error:
                                    value: "\t{\"error\":\"error\"}"
                    description: Something happened
                '404':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/errorResponse'
                            examples:
                                noLoggedTransactions:
                                    value:
                                        error: You have no logged transactions
                    description: You have no logged transactions
            security:
                -
                    Bearer: []
            summary: Get User transaction history
            description: Get User transaction history
        post:
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/postTransactionRequest'
                        examples:
                            createTransaction:
                                value:
                                    accountFrom: 7da3851546037313642
                                    accountTo: 7da3851546037313641
                                    explanation: raha
                                    amount: 50
                required: true
            tags:
                - Transactions
            responses:
                '201':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/successfulTransaction'
                            examples:
                                transactionSuccessful:
                                    value:
                                        message: Transaction created successfully
                    description: Transaction created
                '400':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/errorResponse'
                            examples:
                                invalidAmount:
                                    value:
                                        error: Invalid amount
                                invalidAccountTo:
                                    value:
                                        error: Invalid accountTo
                    description: Invalid amount
                '401':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/errorResponse'
                            examples:
                                missingAuthorizationHeader:
                                    value:
                                        error: Missing Authorization header
                    description: Missing token
                '402':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/errorResponse'
                            examples:
                                insufficentFunds:
                                    value:
                                        error: Insufficient funds
                    description: Insufficent funds
                '404':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/errorResponse'
                            examples:
                                accountNotFound:
                                    value:
                                        error: 'Account not found '
                    description: Account Not Found
            security:
                -
                    Bearer: []
            summary: Make a transaction
            description: Transactions API
    /api/transactions/b2b:
        post:
            tags:
                - Transactions
            responses:
                '200':
                    description: Succesfully received transaction
                '400':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/errorResponse'
                            examples:
                                parsingJwtFailed:
                                    value:
                                        error: Parsing JWT payload failed
                                verifyAccountTo:
                                    value:
                                        error: Account not found
                                bankNotFound:
                                    value:
                                        error: >-
                                            The account sending the funds does not belong to a bank registered
                                            in Central Bank
                                cannotVerifyJwksSignature:
                                    value:
                                        error: 'Cannot verify your signature: The jwksUrl of your bank is invalid'
                                invalidSignature:
                                    value:
                                        error: Invalid signature
                    description: Returns the errors specified
                '500':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/errorResponse'
                            examples:
                                centralBankError:
                                    value:
                                        error: Central Bank error
                                missingJwksUrl:
                                    value:
                                        error: 'Cannot verify your signature: The jwksUrl of your bank is missing'
                    description: Returns the errors specified
...
components:
    schemas:
...
        errorResponse:
            title: Error response
            description: Generic error response
            required:
                - missingAuthorizationHeader
                - error
                - invalidHeaderFormat
                - invalidToken
                - invalidMongooseToken
                - insufficentFunds
                - invalidAmount
                - invalidAccountTo
                - cannotTransferToYourself
                - noLoggedTransactions
            type: object
            properties:
                error:
                    description: Error response
                    type: string
                missingAuthorizationHeader:
                    description: '{"error":"Missing Authorization header"}'
                    type: string
                invalidHeaderFormat:
                    description: "\t{\"error\":\"Invalid Authorization header format\"}"
                    type: string
                invalidToken:
                    description: '{"error":"Invalid token"}'
                    type: string
                invalidMongooseToken:
                    description: '{"error":"Invalid token"}'
                    type: string
                insufficentFunds:
                    description: '{ "error": ''Insufficient funds'' }'
                    type: string
                invalidAmount:
                    description: '{ "error": ''Invalid amount'' }'
                    type: string
                invalidAccountTo:
                    description: '{ "error": ''Invalid accountTo'' }'
                    type: string
                cannotTransferToYourself:
                    description: '{"error": "You cannot transfer to yourself"}'
                    type: string
                noLoggedTransactions:
                    description: '{ "error": "You have no logged transactions" }'
                    type: string
            example:
                error: error message
...
        successfulResponse:
            description: ''
            required:
                - successful
                - successfulHistory
            type: object
            properties:
                successful:
                    description: Successful response
                    type: string
                successfulHistory:
                    description: "{\r\n    \"transactions\": {\r\n        \"createdAt\": \"2020-11-03T17:23:04.056Z\",\r\n        \"status\": \"completed\",\r\n        \"_id\": \"5fa19248f1ff7c46a80eeb5d\",\r\n        \"userId\": \"5fa18d06e11be12268b48733\",\r\n        \"amount\": 500,\r\n        \"currency\": \"EUR\",\r\n        \"accountFrom\": \"f896138686071081018\",\r\n        \"accountTo\": \"f896138686071081017\",\r\n        \"explanation\": \"raha\",\r\n        \"senderName\": \"maarek\",\r\n        \"__v\": 0,\r\n        \"receiverName\": \"maarek1\",\r\n        \"statusDetail\": \"\"\r\n    }\r\n}"
                    type: string
            example:
                name: Ojarand Karl Galkin Hannes-Gregor
                username: Hannes
                account:
                    -
                        account_number: EE2650749058303131
                        balance: 50
                        currency: EUR
        postTransactionRequest:
            description: ''
            required:
                - accountTo
                - amount
                - explanation
                - accountFrom
            type: object
            properties:
                accountTo:
                    description: Reciever
                    type: string
                explanation:
                    description: Transaction explanation
                    type: string
                amount:
                    description: Fund amount
                    type: integer
                accountFrom:
                    description: Sender
                    type: string
            example:
                accountFrom: 7da3851546037313642
                accountTo: 7da3851546037313641
                explanation: raha
                amount: 50
        successfulTransaction:
            title: Root Type for successfulTransaction
            description: ''
            type: object
            properties:
                message:
                    description: Success message
                    type: string
            example:
                message: Transaction created successfully
...
    securitySchemes:
        Bearer:
            scheme: bearer
            type: http